refactor: page와 form 분리
